requirements.txt
alpaca-py==0.38.0
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
DateTime==5.5
Deprecated==1.2.18
dnspython==2.7.0
fastapi==0.115.8
fonttools==4.56.0
frozendict==2.4.6
h11==0.14.0
hopcroftkarp==1.2.5
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
motor==3.7.0
msgpack==1.1.0
multitasking==0.0.11
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
peewee==3.17.9
persim==0.3.7
pillow==11.1.0
platformdirs==4.3.6
polygon-api-client==1.14.3
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyWavelets==1.8.0
requests==2.32.3
ripser==0.6.10
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sseclient-py==1.8.0
starlette==0.45.3
stats==0.1.2a0
ta-lib==0.6.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
wrapt==1.17.2
yfinance==0.2.53
zope.interface==7.2
